feat(TagManager): move scripts to client folder
