+++ src/Kentico.Xperience.TagManager/Admin/Client/.babelrc
