+++ src/Kentico.Xperience.TagManager/Admin/Client/package.json
