+++ src/Kentico.Xperience.TagManager/Admin/Client/src/index.js
