+++ src/Kentico.Xperience.TagManager/Admin/wwwroot/Scripts/ktc-tagmanager.js
@@ -1,0 +1,1 @@
+var e,t;const n="HeadTop",o="HeadBottom",a="BodyTop",c="BodyBottom";null!==(t=(e=window).xperience)&&void 0!==t||(e.xperience={}),window.xperience.tagManager={updateCodeSnippets:async function(){const e=[...document.querySelectorAll("[data-snippet-id]")],t=new Set(e.map((e=>parseInt(e.dataset.snippetId)))),d=await async function(){const e=await fetch("".concat("/kentico.tagmanager","/GetConsentedTags"),{method:"post",headers:{Accept:"application/json","Content-Type":"application/json"}});if(!e.ok)throw new Error(e.statusText);const t=await e.json();return t}();function r(e,t){const n=document.createElement("template");n.innerHTML=e;for(let e of[...n.content.children])if("SCRIPT"===e.nodeName){t(i(e))}else t(e)}function i(e){const t=document.createElement("script");for(let n of[...e.attributes])t.setAttribute(n.name,n.value);return t.text=e.text,t}!function(){const e=document.createElement("template");d.forEach((t=>e.append(t.code)));var i=d.filter((e=>!t.has(e.id)));if(i)for(let e of i){let t;switch(e.location){case n:t=e=>document.head.prepend(e);break;case o:t=e=>document.head.append(e);break;case a:t=e=>document.body.prepend(e);break;case c:t=e=>document.body.append(e)}r(e.code,(e=>t(e)))}}(),function(){for(let t of e.filter((e=>!d.some((t=>t.id===parseInt(e.dataset.snippetId))))))t.remove()}()}};
