+++ src/Kentico.Xperience.TagManager/wwwroot/Scripts/ktc-tagmanager.js.LICENSE.txt
@@ -1,0 +1,1 @@
+/*! regenerator-runtime -- Copyright (c) 2014-present, Facebook, Inc. -- license (MIT): https://github.com/facebook/regenerator/blob/main/LICENSE */
